Update return type to match new prototypes added to sys_public.h in 850981b84bee6aa3bb85f6327d1112413e4b3235

+++ neo/sys/posix/posix_main.cpp
@@ -385,7 +385,7 @@ Sys_DLL_Load
 TODO: OSX - use the native API instead? NSModule
 =================
 */
-int Sys_DLL_Load( const char *path ) {
+size_t Sys_DLL_Load( const char *path ) {
 	void *handle = dlopen( path, RTLD_NOW );
 	if ( !handle ) {
 		Sys_Printf( "dlopen '%s' failed: %s\n", path, dlerror() );
